--- CRMticketSidebar/src/App.jsx
@@ -1,17 +1,0 @@
-
-import Router from './Router'
-import './App.css'
-import Sidebar from './components/sidebar/Sidebar'
-
-
-const App=()=>{
-
-  return(
-    <>
-    <Router/>
-    </>
-  )
-}
-
-
-export default App
