--- CRMticketSidebar/src/index.css
@@ -1,10 +1,0 @@
-*{
-
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-body{
-   font-family: Montserrat, arial,'Times New Roman', Times, serif;
-}
