+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "crm-ticketing",
+  "name": "navigation",
   "private": true,
   "version": "0.0.0",
   "type": "module",
